update token select, fix tx links
